+++ src/SFA.DAS.EmployerIncentives.Api.AcceptanceTests/Steps/ResumePausedPaymentsSteps.cs
@@ -172,7 +172,7 @@ namespace SFA.DAS.EmployerIncentives.Api.AcceptanceTests.Steps
         [Then(@"the requester is informed the apprenticeship incentives have resumed")]        
         public async Task ThenTheRequesterIsInformedTheApprenticeshipIncentiveHasResumed()
         {
-            _response.StatusCode.Should().Be((int)HttpStatusCode.OK);
+            _response.StatusCode.Should().Be(HttpStatusCode.OK);
             var content = await _response.Content.ReadAsStringAsync();
             JsonConvert.SerializeObject(content).Should().Contain("Payments have been successfully Resumed");
         }
@@ -180,7 +180,7 @@ namespace SFA.DAS.EmployerIncentives.Api.AcceptanceTests.Steps
         [Then(@"the requester is informed the apprenticeship incentive is not paused")]
         public async Task ThenTheRequesterIsInformedTheApprenticeshipIncentiveIsNotPaused()
         {
-            _response.StatusCode.Should().Be((int)HttpStatusCode.BadRequest);
+            _response.StatusCode.Should().Be(HttpStatusCode.BadRequest);
             var content = await _response.Content.ReadAsStringAsync();
             JsonConvert.SerializeObject(content).Should().Contain("Payments are not paused");
         }
@@ -188,7 +188,7 @@ namespace SFA.DAS.EmployerIncentives.Api.AcceptanceTests.Steps
         [Then(@"the requester is informed the request is invalid")]
         public async Task ThenTheRequesterIsInformedTheRequestIsInvalid()
         {
-            _response.StatusCode.Should().Be((int)HttpStatusCode.BadRequest);
+            _response.StatusCode.Should().Be(HttpStatusCode.BadRequest);
             var content = await _response.Content.ReadAsStringAsync();
             JsonConvert.SerializeObject(content).Should().Contain("not set");
         }
